Added NL card creation

+++ src/main/java/com/example/trellite/dto/AssignTaskDTO.java
@@ -1,6 +1,7 @@
 package com.example.trellite.dto;
 
+/** A null assigneeId is meaningful here: it clears the card's assignee. */
 public record AssignTaskDTO(
-        Integer assigneeId
+        Long assigneeId
 ) {
 }
